% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdar.R
\name{bc_ratio}
\alias{bc_ratio}
\title{Benefit Cost Ratio}
\usage{
bc_ratio(cif_t0, cif, cif_times, cof_t0, cof, cof_times, r)
}
\arguments{
\item{cif_t0}{cash inflow at time 0}

\item{cif}{cash inflow vector}

\item{cif_times}{cash inflow time vector}

\item{cof_t0}{cash outflow at time 0}

\item{cof}{cash outflow vector}

\item{cof_times}{cash out flow time vector}

\item{r}{nominal interest rate}
}
\value{
Returns BCR
}
\description{
This R function calculates BCR for the given information.
}
\examples{
bc_ratio(cif_t0 = 10, cif = 9,cif_times = 2,cof_t0 = -10, cof = -5, cof_times = 2, r= 0.1)
}
