% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{BinomialTree}
\alias{BinomialTree}
\title{Real Options Pricing using Binomial Lattice}
\usage{
BinomialTree(S, I, Time, r, sigma, dt)
}
\arguments{
\item{S}{State Variable}

\item{I}{Investment}

\item{Time}{Investment Horizon (yearly)}

\item{r}{Rate of Return}

\item{sigma}{Fluctuations in the Price of State Variable}

\item{dt}{Time Intervals within a Year}
}
\description{
This R function will provide real option prices for underlying assets (Infrastructure Projects, Buildings, etc) using binomial lattice.
}
\examples{
BinomialTree(S=50, I=50, Time=5, r=0.2, sigma=0.4, dt=1)
}
