% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{BinomialTree_MC}
\alias{BinomialTree_MC}
\title{Binomial Real Options Pricing with Monte Carlo Simulation}
\usage{
BinomialTree_MC(S, I, Time, r, sigma, dt, MC_loops)
}
\arguments{
\item{S}{State Variable}

\item{I}{Investment}

\item{Time}{Investment Horizon (yearly)}

\item{r}{Rate of Return}

\item{sigma}{Fluctuations in the Price of State Variable}

\item{dt}{Time Intervals within a Year}

\item{MC_loops}{Number of Monte Carlo Simulations}
}
\description{
This R function will provide real option prices and the probability of investment within the investment horizon using binomial lattice and monte carlo simulations.
}
\examples{
BinomialTree_MC(S=10, I=100, Time=10, r=.01, sigma=0.6, dt=1, MC_loops = 1000)
}
