% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccci}
\alias{ccci}
\title{California Construction Cost Index (CCCI)}
\format{
A vector with 168 datapoints.
\describe{
  \item{ccci}{Building Cost Index (BCI) average for San Francisco and Los Angeles, Index}
}
}
\source{
\url{https://www.dgs.ca.gov/RESD/Resources/Page-Content/Real-Estate-Services-Division-Resources-List-Folder/DGS-California-Construction-Cost-Index-CCCI#accordion-68cd4791-da8a-422f-b329-4aa358dec53c}
}
\usage{
ccci
}
\description{
CCCI is a monthly index developed based upon Building Cost Index (BCI) for San Francisco and Los Angeles.
This time series represent monthly ccci for the period of 14 years from January 2007 to November 2018.
}
\keyword{datasets}
