% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdar.R
\name{cfd}
\alias{cfd}
\title{Cashflow Diagram}
\usage{
cfd(cf_t0, cf, times)
}
\arguments{
\item{cf_t0}{Cashflows at time 0}

\item{cf}{Cashflow vector}

\item{times}{Cashflow times}
}
\value{
Returns the Cashflow digram for the given cashflow streams
}
\description{
This R function will provide a cashflow diagram with the given cashflow streams and their times.
}
\examples{
cfd(cf_t0 = -1000, cf=c(200,200,1000), times = c(1,1,2))
}
