% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop}
\alias{cop}
\title{US Crude Oil Price (COP)}
\format{
A vector with 192 datapoints.
\describe{
  \item{cop}{US Crude Oil Price, Dollar/Barrel}
}
}
\source{
\url{https://tradingeconomics.com/commodity/crude-oil}
}
\usage{
cop
}
\description{
A monthly Crude Oil Price
}
\keyword{datasets}
