% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{United States Consumer Price Index (CPI)}
\format{
A vector with 192 datapoints.
\describe{
  \item{cpi}{U.S.Consumer Price Index (CPI), Index}
}
}
\source{
\url{https://tradingeconomics.com/united-states/consumer-price-index-cpi}
}
\usage{
cpi
}
\description{
In the United States, the Consumer Price Index or CPI measures changes in the prices paid by consumers for a basket of goods and services.
This time series represents monthly U.S.  Consumer Price Index (CPI) from 2003 to 2018.
}
\keyword{datasets}
