% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R, R/cdar.R
\name{dpbp_ia}
\alias{dpbp_ia}
\title{Discounted payback period for multiple options}
\usage{
dpbp_ia(project1 = NA, project2 = NA, project3 = NA, r, cf_t0)

dpbp_ia(project1 = NA, project2 = NA, project3 = NA, r, cf_t0)
}
\arguments{
\item{project1}{cashflow for project 1}

\item{project2}{cashflow for project 2}

\item{project3}{cashflow for project 3}

\item{r}{nominal interest rate}

\item{cf_t0}{cashflow at time 0, True if projects' cashflows vector include initial cashflow or False vice versa.}
}
\value{
Returns Discounted payback period for multiple projects

Returns Discounted payback period for multiple projects
}
\description{
This R function calculates payback period for up to 3 different options.

This R function calculates payback period for up to 3 different options.
}
\examples{
dpbp_ia(project1 = c(-20, 9),project2 = c(-10, 10), project3 = c(-10, 9),cf_t0 = TRUE, r = 0.1)
dpbp_ia(project1 = c(-20, 9),project2 = c(-10, 10), project3 = c(-10, 9),cf_t0 = TRUE, r = 0.1)
}
