% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elc}
\alias{elc}
\title{U.S. Employment Level in Construction (ELC)}
\format{
A vector with 192 datapoints.
\describe{
  \item{elc}{Employment Level in Construction (ELC), Thousands}
}
}
\source{
\url{https://www.bls.gov/iag/tgs/iag23.htm}
}
\usage{
elc
}
\description{
A monthly index for Employment Level in Construction industry.
This time series represents employment level in construction industry (thousands) from 2003 to 2018.
}
\keyword{datasets}
