% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdar.R
\name{euv}
\alias{euv}
\title{Equalent Uniform Value or Annuity}
\usage{
euv(
  pv = NA,
  fv = NA,
  n = NA,
  a = NA,
  q = NA,
  r = NA,
  ic = 1,
  pf = 1,
  imm = TRUE,
  plot = FALSE
)
}
\arguments{
\item{pv}{present value}

\item{fv}{future value}

\item{n}{number of periods}

\item{a}{amount of the first payment}

\item{q}{payment increment amount per period}

\item{r}{nominal interest rate convertible ic times per period}

\item{ic}{interest conversion frequency per period}

\item{pf}{Payment frequency - number of payments per year}

\item{imm}{option for annuity immediate or annuity due, default is immediate (TRUE)}

\item{plot}{option to display a time diagram of the payments}
}
\value{
Returns Annuity
}
\description{
This R function calculate annuity for the present value, future value, number of payments/periods, amount of the first payment, the payment increment amount per period, and/or the interest rate for an arithmetically growing annuity.
}
\examples{
euv(pv=1000, fv=NA, n=10, a=NA, q=0, r=.05, ic=1, pf=1, imm=FALSE)
}
