% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{U.S. Gross Domestic Product (GDP)}
\format{
A vector with 66 datapoints.
\describe{
  \item{gdp}{U.S.Gross Domestic Product (GDP), Billion Dollars}
}
}
\source{
\url{https://fred.stlouisfed.org/series/GDP}
}
\usage{
gdp
}
\description{
Gross domestic product (GDP), the featured measure of U.S. output, is the market value of the goods and services produced by labor and property located in the United States.For more information, see the Guide to the National Income and Product Accounts of the United States (NIPA) and the Bureau of Economic Analysis.
This time series represents U.S. GDP (Billions of Dollars) from first quarter of 2003 to second quarter of 2019.
}
\keyword{datasets}
