% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcs}
\alias{hcs}
\title{Highway Construction Spending (HCS)}
\format{
A vector with 198 datapoints.
\describe{
  \item{hcs}{Highway Construction Spending, in US billion dollars}
}
}
\source{
\url{https://www.census.gov/construction/c30/historical_data.html}
}
\usage{
hcs
}
\description{
Value of federal construction put in place - highway and street is published by U.S. Census Bureau
This time series is a monthly measure of the value of construction spending on highways and street by federal government.
This time series represent monthly hcs for the period of 16 years from 2003 to 2018.
}
\keyword{datasets}
