% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ihc}
\alias{ihc}
\title{Iowa Highway Construction (IHC)}
\format{
A vector with 48 datapoints.
\describe{
  \item{ihc}{Changes in the prices of highway construction costs over time}
}
}
\source{
\url{https://iowadot.gov/}
}
\usage{
ihc
}
\description{
A quarterly index published by IOWA Department of Transportation (IOWADOT).
This index is composed of three construction categories (six items), including roadway excavation, surfacing, and structures, on awarded contracts. IHC is calculated from the 1987 BASE to measure the average changes in the prices of highway construction costs over time.
This time series represent quarterly ihc for the period of 12 years from 2007 to 2018.
}
\keyword{datasets}
