% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{investment_assessment}
\alias{investment_assessment}
\title{Investment assessment for multiple options}
\usage{
investment_assessment(project1 = NA, project2 = NA, project3 = NA, r, cf_t0)
}
\arguments{
\item{project1}{cashflow for project 1}

\item{project2}{cashflow for project 2}

\item{project3}{cashflow for project 3}

\item{r}{nominal interest rate}

\item{cf_t0}{cashflow at time 0, True if projects' cashflows vector include initial cashflow or False vice versa.}
}
\value{
Returns NPV, EUV, IRR, and discounted payback period for multiple projects
}
\description{
This R function calculates NPV, EUV, IRR, and payback period for up to 3 different options.
}
\examples{
investment_assessment(project1 = c(-2000, -1250, 1000, 1000, 2000, 2000),project2 = c(-1000, 1000, 1000, 10, 200, 2000), project3 = c(-1000, 1250, 100, 1000, 2000, 20),cf_t0 = TRUE, r = 0.1)
}
