% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{irr_ia}
\alias{irr_ia}
\title{Internal Rate of Return for multiple options}
\usage{
irr_ia(project1 = NA, project2 = NA, project3 = NA, cf_t0)
}
\arguments{
\item{project1}{cashflow for project 1}

\item{project2}{cashflow for project 2}

\item{project3}{cashflow for project 3}

\item{cf_t0}{cashflow at time 0, True if projects' cashflows vector include initial cashflow or False vice versa.}
}
\value{
Returns IRR for multiple projects
}
\description{
This R function calculates IRR for up to 3 different options.
}
\examples{
irr_ia(project1 = c(-20, 9),project2 = c(-10, 10), project3 = c(-10, 9),cf_t0 = TRUE)
}
