% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{lcca_montecarlo}
\alias{lcca_montecarlo}
\title{Probabilistic Life Cycle Cost Analysis}
\usage{
lcca_montecarlo(
  comp1 = NA,
  comp2 = NA,
  comp3 = NA,
  comp4 = NA,
  comp5 = NA,
  recurring_cost = NA,
  r,
  n,
  n_loop
)
}
\arguments{
\item{comp1}{A vector for the first initial cost component  #comp1 = rtriangle(n_loop, a =10, b = 200, c = 30)}

\item{comp2}{A vector for the second initial cost component #comp2 = rnorm(n_loop, mean=10, sd=3)}

\item{comp3}{A vector for the third initial cost component  #comp3 = rlnorm(n_loop, meanlog = 10, sdlog = 1)}

\item{comp4}{A vector for the forth initial cost component  #comp4 = runif(n_loop, min = 0, max = 25)}

\item{recurring_cost}{The annual reccuring cost of the project}

\item{r}{Rate of Return}

\item{n}{Investment Horizon (yearly)}

\item{n_loop}{Number of Monte Carlo Iterations}
}
\description{
This R function will provide probabilistic LCCA for a projects with several uncertain initail cost components.
}
\examples{
lcca_montecarlo(comp1 = comp1, comp2 = comp2, comp3 = comp3, comp4 = comp4, r=0.05, n_loop=n_loop, n=50, recurring_cost = 100)
}
