% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ms}
\alias{ms}
\title{Money Supply (M1) (MS-M1)}
\format{
A vector with 192 datapoints.
\describe{
  \item{ms}{US Money Supply (M1), Billion Dollars}
}
}
\source{
\url{https://tradingeconomics.com/united-states/money-supply-m1}
}
\usage{
ms
}
\description{
M1 is the money supply that is composed of physical currency and coin, demand deposits, travelers' checks, other checkable deposits, and negotiable order of withdrawal (NOW) accounts. M1 includes the most liquid portions of the money supply because it contains currency and assets that either are or can be quickly converted to cash.
This time series represents US Money Supply (M1) from 2003 to 2018.
}
\keyword{datasets}
