% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhcci}
\alias{nhcci}
\title{National Highway Construction Cost Index (NHCCI)}
\format{
A vector with 64 datapoints.
\describe{
  \item{nhcci}{National Highway Construction Cost Index (NHCCI), Cost Index}
}
}
\source{
\url{https://www.fhwa.dot.gov/programadmin/pricetrends.cfm}
}
\usage{
nhcci
}
\description{
A quarterly price index published by Federal Highway Administration (FHWA) to measure the average changes in the prices of highway construction costs over time.
This time series represent quarterly price index nhcci for the period of 16 years from 2003 to 2018.
}
\keyword{datasets}
