% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{npv}
\alias{npv}
\title{Net Present Value}
\usage{
npv(cf_t0, cf, times, r)
}
\arguments{
\item{cf_t0}{Cashflows at time 0}

\item{cf}{Cashflow vector}

\item{times}{Cashflow times}

\item{r}{rate of return}
}
\value{
Returns the NPV for the given cashflow streams
}
\description{
This R function will calculate NPV for the given cashflow streams.
}
\examples{
npv(cf_t0 = -1000 , cf = c(100,300)  , times = c(3,2)  , r = 0.05)
}
