% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdar.R
\name{npv_ia}
\alias{npv_ia}
\title{Net Present Value for multiple options}
\usage{
npv_ia(project1 = NA, project2 = NA, project3 = NA, r, cf_t0)
}
\arguments{
\item{project1}{cashflow for project 1}

\item{project2}{cashflow for project 2}

\item{project3}{cashflow for project 3}

\item{r}{nominal interest rate convertible ic times per period}

\item{cf_t0}{cashflow at time 0, True if projects' cashflows vector include initial cashflow or False vice versa.}
}
\value{
Returns NPV for multiple projects
}
\description{
This R function calculate NPV for up to 3 different options.
}
\examples{
npv_ia(project1 = c(-20, 9),project2 = c(-10, 10), project3 = c(-10, 9),r = 0.05,cf_t0 = TRUE)
}
