% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdar.R
\name{tvm}
\alias{tvm}
\title{Time Value of Money}
\usage{
tvm(pv = NA, fv = NA, n = NA, r = NA, ic = 1)
}
\arguments{
\item{pv}{present value}

\item{fv}{future value}

\item{n}{number of periods}

\item{r}{nominal interest rate convertible ic times per period}

\item{ic}{interest conversion frequency per period}
}
\value{
Returns the NPV or FV for the given information
}
\description{
This R function can calculate present value and future value the given information.
}
\examples{
tvm(pv=50, n=5, r=.04)
}
