% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{BinomialTree}
\alias{BinomialTree}
\title{Real Options Pricing using Binomial Lattice}
\usage{
BinomialTree(S, I, Time, r, sigma, n)
}
\arguments{
\item{S}{State Variable}

\item{I}{Investment}

\item{Time}{Time Intervals within a Year}

\item{r}{Rate of Return}

\item{sigma}{Fluctuations in the Price of State Variable}

\item{n}{Investment Horizon (yearly)}
}
\description{
This R function will provide real option prices for underlying assets (Infrastructure Projects, Buildings, etc) using binomial lattice.
}
\examples{
BinomialTree(S=60, I=50,Time=1, r=.1,sigma=0.2,n=5)
}
