% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{BinomialTree}
\alias{BinomialTree}
\title{Real Options Pricing using Binomial Lattice}
\usage{
BinomialTree(S, I, Time, r, sigma, dt, k = NA)
}
\arguments{
\item{S}{State Variable}

\item{I}{Investment}

\item{Time}{Investment Horizon (yearly)}

\item{r}{Rate of Return}

\item{sigma}{Fluctuations in the Price of State Variable}

\item{dt}{Time Intervals within a Year}

\item{k}{Risk-Adjusted Growth Factor}
}
\value{
Returns a binomial tree for the state variable "S", cashflow matrix calculated from the binomial tree and the investment cost, decision matrix for investment for different situations through the investment horizon, and a binomial tree plot.
}
\description{
This R function will provide real option prices for underlying assets (Infrastructure Projects, Buildings, etc) using binomial lattice.
}
\examples{
BinomialTree(S=50, I=30, Time=5, r=0.07, sigma=0.15, dt=1, k =1.02)
}
