% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R, R/cdar.R
\name{BinomialTree_MC}
\alias{BinomialTree_MC}
\title{Binomial Real Options Pricing with Monte Carlo Simulation}
\usage{
BinomialTree_MC(S, I, Time, r, sigma, dt, k = NA, imm = TRUE, MC_loops)

BinomialTree_MC(S, I, Time, r, sigma, dt, k = NA, imm = TRUE, MC_loops)
}
\arguments{
\item{S}{State Variable}

\item{I}{Investment}

\item{Time}{Investment Horizon (yearly)}

\item{r}{Rate of Return}

\item{sigma}{Fluctuations in the Price of State Variable}

\item{dt}{Time Intervals within a Year}

\item{k}{Risk-Adjusted Growth Factor}

\item{imm}{Cashflows are collected immediately after investemnt in the same yesr if it is TRUE, otherwise collect cashflows after one year (default is TRUE)}

\item{MC_loops}{Number of Monte Carlo Simulations}
}
\value{
Returns a binomial tree for the state variable "S", cashflow matrix calculated from the binomial tree and the investment cost, decision matrix for investment for different situations through the investment horizon, a binomial tree plot, and the likelihood of implementation plot.

Returns a binomial tree for the state variable "S", cashflow matrix calculated from the binomial tree and the investment cost, decision matrix for investment for different situations through the investment horizon, a binomial tree plot, and the likelihood of implementation plot.
}
\description{
This R function will provide real option prices and the probability of investment within the investment horizon using binomial lattice and monte carlo simulations.

This R function will provide real option prices and the probability of investment within the investment horizon using binomial lattice and monte carlo simulations.
}
\examples{
BinomialTree_MC(S=10, I=100, Time=10, r=.01, sigma=0.6, dt=1, imm = TRUE, MC_loops = 1000)
BinomialTree_MC(S=10, I=100, Time=10, r=.01, sigma=0.6, dt=1, imm = TRUE, MC_loops = 1000)
}
