% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructionAnalytics.R
\name{Lcca_MonteCarlo}
\alias{Lcca_MonteCarlo}
\title{Probabilistic Life Cycle Cost Analysis}
\usage{
Lcca_MonteCarlo(
  comp1 = NA,
  comp2 = NA,
  comp3 = NA,
  comp4 = NA,
  comp5 = NA,
  recurring_cost = NA,
  r,
  n,
  n_loop
)
}
\arguments{
\item{comp1}{A vector for the first initial cost component  #Example for comp1 = rtriangle(n_loop, a =10, b = 200, c = 30)}

\item{comp2}{A vector for the second initial cost component #Example for comp2 = rnorm(n_loop, mean=10, sd=3)}

\item{comp3}{A vector for the third initial cost component  #Example for comp3 = rlnorm(n_loop, meanlog = 10, sdlog = 1)}

\item{comp4}{A vector for the forth initial cost component  #Example for comp4 = runif(n_loop, min = 0, max = 25)}

\item{comp5}{A vector for the fifth initial cost component  #Example for comp5 = runif(n_loop, min = 0, max = 25)}

\item{recurring_cost}{The annual reccuring cost of the project}

\item{r}{Rate of Return}

\item{n}{Investment Horizon (yearly)}

\item{n_loop}{Number of Monte Carlo Iterations}
}
\description{
This R function will provide probabilistic LCCA for a projects with several uncertain initail cost components.
}
\examples{
Lcca_MonteCarlo(comp1 = rnorm(1000, mean=34, sd=8), comp2 = rnorm(1000, mean=10, sd=3),
comp3 = rlnorm(1000, meanlog = 10, sdlog = 1), r =0.05, recurring_cost = 100,  n=10, n_loop = 100 )
}
